Disabled `RemoveRelationCopiesTransformer`.
This optimization removes relations that are defined with a single rule
of the form R(x, y, ...) :- S(x, y, ...), i.e., are a direct copy of
another relation. Unfortunately, it does not seem to handle the case
where R *also* is read from an input file, and so is actually a superset
of S. This is a quick fix until we can properly verify and fix this bug.