doc comments and annotations in Souffle datalog
Annotations are free meta-data that can be attached to most items of a
Souffle datalog program.

A documentation comment is a syntactic sugar for a `doc` annotation.

The design is close to Rust's *attributes*.

All program's and component's items can have *outer* annotations of
the form `@[annotation]` or `/// doc comment`. Attributes and ADT
branches also accept outer annotations.

Some items can have *inner* annotations of the for `@![annotation]`
or `//! doc comment`:
- after `{` of a component, before the first component's item.
- after `{` of an ADT branch, before the first attribute of the branch.
- after `:-` of a clause, before the first constraint.

Annotations should not appear before `.include` or before any `}` or
before the end of a file.

An annotation's payload may be of these forms:
- an identifier: `@[IDENT]`
- possibly followed by delimited token stream: `@[IDENT DELIM]`
- or `=` and a delimited token stream: `@[IDENT = DELIM]` or
  `@[IDENT = TOKEN TT*]`

Where:
- `IDENT` is an identifier or a keyword of Souffle.
- `TOKEN` is any token except delimiters.
- `TS ::= TT+` is a token-stream.
- `TT ::= TOKEN | DELIM` is a token-tree.
- `DELIM ::= '(' TS? ')' | '[' TS? ']' | '{' TS? '}'` is a delimited
  token-stream.