Ensure we can read blank lines in CSV files.
There was a problem handling blank lines in a quoted
CSV column when reading RFC4180 mode.
Instead of continuing to read more lines, it dropped out
assuming there was at least 1 character. Now, if the
line we read contains 0 characters, it immediately goes
back to read more lines.

I did not notice this before, because the test did not
use rfc4180 csv files for input or output.