fix `apply` on user-defined aggregate `init`
The `init` expression of user-define aggregates was never
handed-over to `apply`'s `mapper`.

One visible issue is that the `TupleId` transformer would not update the
tuple of the `init` expression while shuffling tuple identifiers.
That would then result in the `init` expression peeking data in the
wrong tuple during execution of the RAM program.