Resolution of `inline` atom in `no_inline` clause
When the clause of a `no_inline` relation contains an atom of an
`inline` relation, the atom is not inlined. Hence the `inline` relation
should appear as a normal relation to support non inlined atoms.

The interpretation is as follows:
- do not inline any atom in clauses of relations marked `no_inline`.
- inline atoms of relations marked `inline`, except in clauses of
  relations marked `no_inline`.

The fix consists in removing the `inline` mark on every relations at the
end of the InlineRelationsTransformer pass.

fix #2318