fix ungrounded error when destructuring with don't-cares
Name unnamed variables that appear in record and branch
inits before the resolution of aliases.

fix 2482

After the `ResolveAliasesTransformer`, all  occurences of `t` are replaced with `[ta,_]`, and the groundedness constraint system cannot give a bounded value to `[ta,_]` in the head (`var([ta,_])->0` in the `Solution`).

```
Clause: pred([ta,_],ta) :-
   fact([ta,_]).
Problem:
{
        var([ta,_]) ⇒ var(ta),
        var([ta,_]) ⇒ var(_),
        var(ta) ∧ var(_) ⇒ var([ta,_]),
        var([ta,_]) is true,
        var([ta,_]) ⇒ var(ta),
        var([ta,_]) ⇒ var(_),
        var(ta) ∧ var(_) ⇒ var([ta,_])
}
Solution:
{var(_)->1,var(_)->0,var([ta,_])->1,var([ta,_])->0,var(ta)->1}
```

When you provide an explicit variable for the second part of the record, the constraint system can be solved because the second member of the record has is bound by the variable (`var(_fix)->1` and `var([ta,_fix])->1`):

```
Clause: pred([ta,_fix],ta) :-
   fact([ta,_fix]).
Problem:
{
        var([ta,_fix]) ⇒ var(ta),
        var([ta,_fix]) ⇒ var(_fix),
        var(ta) ∧ var(_fix) ⇒ var([ta,_fix]),
        var([ta,_fix]) is true,
        var([ta,_fix]) ⇒ var(ta),
        var([ta,_fix]) ⇒ var(_fix),
        var(ta) ∧ var(_fix) ⇒ var([ta,_fix])
}
Solution:
{var([ta,_fix])->1,var([ta,_fix])->1,var(ta)->1,var(_fix)->1}
```

To fix this, we need to transform `_` into unique variables before the `ResolveAliasesTransformer`.