robustify the type-analysis debug report
The type-analysis debug report used to crash on various assertions
because it assumed a well-typed program. However the first type-analysis
runs before the semantics checker, hence the program may be ill-typed at
that time.

With this change the type-analysis debug report printer makes no
assumptions about the expected types, instead it prints the infered
types.