Do not use SelingerProfileSipsMetric if no profile available (#2517)
* Add regression test for issue 2426.

* Do not use SelingerProfileSipsMetric if no profile available.

There is a bug #2426, which causes an assertion in the
SelingerProfileSipsMetric::getReordering.

Rther than fixing the problem I decided to work-around it for
the time being, until we have a proper fix.

We can work around this problem in one of two ways:
1. don't assert and just don't reorder atoms (this should be fine)
2. don't create the metric in the first place, if there is no profile

I decided to use #2 as that seems to be the safer option.

I added code for a warning, but disabled it because it is incompatible
with the ctest suite.

* Disable checking stderr output for scheduler tests.

Printing the warning about --auto-schedule not
being enabled causes a problem with tests.

For now, just disable checking of stderr outputs in the
test. We're still checking for other outputs.