Limit application of SelingerProfileSipsMetric (#2546)
* Added a reproducer Selinger exponential blowup.

I've used 50 atoms in a single
rule which causes the Selinger SIPS metric to basically
run for very long time. Memory also blows up.

* Limit application of SelingerProfileSipsMetric

The SelingerProfileSipsMetric uses an algorithm called
Selinger's algorithm. This algorithm is exponential in
the number of relations (n*2^N). With inlining, the number
of relations for a rule can quickly grow and we need to ensure
things don't blow up.

I've limited the number of atoms to just 16 after writing a reproducer
that allowed me to test different values. I made the parameter
a macro, which could allow users to define a compile time parameter
and change the limit easily.